feat: replace `reaction plugins` usage
Use the scripts now available in this repo, which do the same thing.
