+++ .reaction/scripts/appSetup.mjs
@@ -1,12 +1,9 @@
-// Assumes Node 8.x
-import _ from "lodash";
-import childProcess from "child_process";
 import Log from "./logger";
 import loadPlugins from "./loadPlugins";
 import loadStyles from "./loadStyles";
 import provisionAssets from "./provisionAssets";
 
-function run() {
+export default function appSetup() {
   let start, sec, ns;
 
   start = process.hrtime();
@@ -26,24 +23,4 @@ function run() {
   provisionAssets();
   [sec, ns] = process.hrtime(start);
   Log.info(`Provisioning assets took ${sec}s ${ns / 1000000}ms\n`);
-
-  // Whatever debugging-related command line arguments were passed in to
-  // the first node process, forward them along through meteor
-  const inspect = process.argv
-    .filter((arg) => arg.startsWith("--inspect"))
-    .join(" ");
-  let cmd = `meteor run --no-lint --no-release-check --raw-logs ${inspect}`;
-
-  Log.info(`Running command: ${cmd}`);
-  cmd = `REACTION_METEOR_APP_COMMAND_START_TIME=${Date.now()} ${cmd}`;
-
-  try {
-    childProcess.execSync(cmd, { stdio: "inherit" });
-  } catch (err) {
-    Log.default(err);
-    Log.error("\nError: App failed to start");
-    process.exit(1);
-  }
 }
-
-run();
