+++ .reaction/scripts/build.mjs
@@ -1,0 +1,3 @@
+import appSetup from "./appSetup";
+
+appSetup();
