+++ docker-compose.yml
@@ -21,7 +21,7 @@ services:
     build:
       context: .
       target: meteor-dev
-    command: bash -c "time meteor npm install --no-audit && node ./.reaction/waitForMongo.js && node --experimental-modules ./.reaction/run/index.mjs"
+    command: bash -c "time meteor npm install --no-audit && node ./.reaction/waitForMongo.js && node --experimental-modules ./.reaction/scripts/run.mjs"
     depends_on:
       - mongo
     env_file:
