+++ package.json
@@ -210,11 +210,11 @@
   },
   "scripts": {
     "devserver": "ROOT_URL=http://localhost:3030 MONGO_URL=mongodb://localhost:27017/reaction NODE_ENV=reaction-node BABEL_DISABLE_CACHE=1 nodemon ./imports/node-app/devserver/index.js",
-    "dev": "node --experimental-modules ./.reaction/run/index.mjs",
-    "inspect": "node --experimental-modules ./.reaction/run/index.mjs --inspect",
-    "inspect-brk": "node --experimental-modules  ./.reaction/run/index.mjs --inspect-brk",
-    "inspect-docker": "node --experimental-modules  ./.reaction/run/index.mjs --inspect=0.0.0.0:9229",
-    "inspect-brk-docker": "node --experimental-modules  ./.reaction/run/index.mjs --inspect-brk=0.0.0.0:9229",
+    "dev": "node --experimental-modules ./.reaction/scripts/run.mjs",
+    "inspect": "node --experimental-modules ./.reaction/scripts/run.mjs --inspect",
+    "inspect-brk": "node --experimental-modules  ./.reaction/scripts/run.mjs --inspect-brk",
+    "inspect-docker": "node --experimental-modules  ./.reaction/scripts/run.mjs --inspect=0.0.0.0:9229",
+    "inspect-brk-docker": "node --experimental-modules  ./.reaction/scripts/run.mjs --inspect-brk=0.0.0.0:9229",
     "lint": "eslint .",
     "lint:errors": "eslint . --quiet",
     "test": "npm run test:unit && npm run test:integration && npm run test:app",
